Fix component index: it is int32, not int64 (found via a real recording)
The player probe reported only 3 components (Pathfinding, Animated,
Inventories) and no Render, so the projection proof could not run. This
is the record/replay layer paying off: the 2 KB recording captured the
real bucket, so the cause was traceable offline without the game.

The hash-bucket entry's Index is a 32-BIT int at +0x08 (the AHK
reference reads it as "Int"); bytes +0x0C..0x0F are a separate field.
Reading Index as 64-bit merged the two, so every entry whose following 4
bytes were non-zero decoded to a huge out-of-range index and got
dropped. Only the entries whose high dword happened to be 0 survived -
which is exactly the 3 that appeared. Re-reading the same recording's
indices as int32 gives a clean dense 0..12, all in range.

- schema: ComponentLookupEntry.Index i64 -> i32, with the failure
  documented on the struct.
- EntityReader: reads Index as int32.
- EntityReaderTests: place the index as int32 AND poison the next 4
  bytes with a realistic non-zero value, so an int64 regression fails
  the test.

58 green. The projection proof still needs one more --record: this
recording never read the Render component (its live run skipped it via
the bug), so the player's world position was never captured. With the
fix, the next recording will carry all components + Render + the
projection.

+++ src/PoEformance.Game/Entities/EntityReader.cs
@@ -158,7 +158,9 @@ public sealed class EntityReader
         {
             ulong entryAddress = dataFirst + (ulong)(i * _entrySize);
             ulong namePtr = _reader.ReadPointer(entryAddress + (ulong)_entryNamePtr);
-            long index = _reader.Read<long>(entryAddress + (ulong)_entryIndex);
+            // Index is a 32-bit int; reading 8 bytes here merges it with the next field
+            // and rejects most components (see the schema comment on ComponentLookupEntry).
+            int index = _reader.Read<int>(entryAddress + (ulong)_entryIndex);
 
             // Empty hash slots carry an out-of-range index; skip them.
             if (index < 0 || index >= componentCount || namePtr == 0)
